requirements.txt
Jinja2
PyMySQL
PyYAML
SQLAlchemy
pathlib
pyserial
python-can
python-dotenv
watchdog